인스타 클론코딩 완료!

+++ src/main/java/com/cos/photogramstart/web/dto/subscribe/SubscribeDto.java
@@ -1,5 +1,7 @@
 package com.cos.photogramstart.web.dto.subscribe;
 
+import java.math.BigInteger;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,9 +13,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class SubscribeDto {
+	
 	private int id;
 	private String username;
 	private String profileImageUrl;
-	private Integer subscribeState;
-	private Integer equalUserState;
+	private BigInteger subscribeState;
+	private BigInteger equalUserState;
+	
 }
